feat(partida): adiciona BuscarPartidaService e busca de partida por id
- Corrige PartidaRepository para aceitar Long em JpaRepository<Partida, Long>.
- Cria exceção PartidaNaoEncontradaException e adiciona seu tratamento no GlobalApiExceptionHandler.
- Adiciona novo código em ErroCodigo para essa situação.
- Implementa método de busca por id no controller de partida.
- Atualiza checklist dos requisitos 7 a 9 de partida e adiciona novo campo de comentario em anotacoes.md.

+++ src/main/java/br/com/meli/teamcubation_partidas_de_futebol/partida/repository/PartidaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-public interface PartidaRepository extends JpaRepository<Partida, Integer> {
+public interface PartidaRepository extends JpaRepository<Partida, Long> {
     @Query(value = """
     SELECT CASE WHEN COUNT(*) > 0 THEN 1 ELSE 0 END
     FROM partida
